Skip writing to storage unchanged interfaces
The default interface updater receives periodic updates with the
current interfaces, computes the difference between the set of
previous interfaces and the current ones, and if any are new,
have changed or have been deleted, are updated accordingly in the
backend storage (ZooKeeper).

This fixes an issue in the default interface updater, where
unchanged interfaces are always updated in storage even if the
interface has not changed, because the code was not comparing
the old and new interfaces. Therefore, this caused a new write to
ZooKeeper for every interface update (approximately 1 per second).

Change-Id: I1ac94876f0a1dd337d1cb7a5c65a0ae622637d17
Fixes: MI-1048
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>