FlowState: file transfer between agents
The new way to transfer flowstate between agents will need agents to
receive a flow state file stream from a different agent and then
transfering the corresponding flowstate back to an interested agent.

For legacy support and migrations, we will continue writing to Cassandra
and add a flag in the configuration to enable this behaviour.

Ref: MI-987

Change-Id: I774a6ebbca3d33475895204c1adba6ecc3645ae0
Signed-off-by: Mateo Hermida <mateo@midokura.com>