10kp: State table storage returns table arguments
State tables for an object  may take an optional sequence of
arguments (e.g. bridge MAC-port tables take an argument which is
the VLAN ID). Sometimes, we need to retrive data from all state
tables that belong to a specific object, for example in the API
to return the all entries found.

This patch adds a new method that returns the found state table
arguments for a given object and optional argument list.

Ref: MI-1004

Change-Id: Icb4dc3ce89ee82f3b0e9bf5f2c5d56f3407bf0c4
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>