MDTS: remove mariadb and point to neutron v1
We were sharing the same neutron container for v1.9 and v5 flavours.
However, the db schema and configuration is different in both
architectures, failing to bring up the neutron container in v1.9.

There's now a v1.9 based neutron container on midonet-sandbox. This
patch removes the now obsolete midonet.ini files (as they are part
of the container) and point to the new container within the flavours.

Change-Id: I22c2ebed7a9edad28f1fd400074a7feeaa308bcc
Signed-off-by: Xavi León <xavi.leon@midokura.com>