Add agent minion to handle local port binding requests
Unlike the rest api on the cluster, this one listens on
a unix domain socket in order to simplify credential management.

Note: the code under hacks directory are necessary to use our
UnixDomainSocketChannel with jetty-9.3.x, which assumes
ServerSocketChannel/SocketChannel.  Hopefully we can get rid of
most of them with jetty-9.4.x, which will have unix domain socket
support. [1]

[1] https://webtide.com/unix-sockets-for-jetty-9-4/

Ref: IN-155
Signed-off-by: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: I336b7505f124a6dcff9425a227b66c5b813150b5