MDTS: upgrade docker-py to make it compatible with py3
Midonet containers integrations (raven, kuryr) aim to be
python3 compatible so they require midonet-sandbox to install a newer
version of docker-compose which installs a newer version of the python
docker interface. Upgrade our requirement to be compatible with the new
midonet-sandbox so we don't downgrade docker-py if using the same
virtual environment.

Change-Id: If32f69efcae2dd1bb7b258bdd4789fd3946b95d3
Signed-off-by: Xavi León <xavi.leon@midokura.com>