WaitingRoom handles reentry of objects correctly
There was a bug in the waiting room, that if a packet context was postponed
once, then completed and reused, and the subsequent reuse was also
postponed, the timeout from the first usage would trigger the drop of
the context, as we don't remove the timeout itself, but rather the
reference to the object in a hash set.

This is fixed by turning the hash set in waiting room into a hash map,
and stores the timeout for the entry in the hash map. On expiry, the
timeout from the expiry queue is compared to the entry in the hash map
to ensure that the correct context is being expired.

Change-Id: Ib969e7f401f43b5f52d1b53900e95d14bdedc5f8
Signed-off-by: Ivan Kelly <ivan@midokura.com>