FlowState: Add config values for new handlers
The new flow state transfer handlers will have and additional TCP server
listening for transfer request. We change the name of the original UDP
port and add the new one. We also need new flags for whether we are
reading/writing to the legacy Cassandra cluster or not. At some point we
shouldn't be using it at all any more, but we need to be able to switch
easily.

Ref: MI-987

Change-Id: I8328c30f6d3eebd5320ba9df591e92512679b68d
Signed-off-by: Mateo Hermida <mateo@midokura.com>