FlowState: internal mappings to ProtoBuf objects
Adds internals objects that can be mapped directly to the Flow State
Protocol ProtoBuf messages that are exchanged in a transfer. This will
ease the work on the more cumbersome auto-generated objects and help
with the creation of the ProtoBuf objects.

A builder and parser objects are also added to avoid the boilerplate of
the translations.

Ref: MI-987

Change-Id: I7b526266dcc152d08649d5df58c81d909d9bfc8a
Signed-off-by: Mateo Hermida <mateo@midokura.com>