FlowState: TCP client for transfer requests
Adds a TCP client that can perform a full flow request transfer on the
client Midonet Agent. It uses the builder and parser utilities to build
a request message and listens on the same TCP socket for the flowstate
response from the Midonet Agent who used to own the port.

Ref: MI-987

Change-Id: Ifb979e43de8aa1eeb86223c4f855f22e99a2134d
Signed-off-by: Mateo Hermida <mateo@midokura.com>