FlowState: introduce byte buffer block writer and reader
This patch introduces an input and output streams that use a ring
buffer of byte buffers (blocks) as the underlying storage mechanism.
Blocks are added on-demand to the ring buffer. Each block contains
a header specifying at least the length of the data written.
This patch also introduces an expiration mechanism for the
blocks allocated to a given ring buffer. After a specified amount of time,
blocks are removed from the tail of the ring buffer and the header
cleared. This patch slightly extends the ring buffer abstraction,
including documentation and unit testing for the newly added
functionality. This components will be used in subsequent patches to
implement a memory mapped ring buffer to store flow state.

Ref.: MI-985
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: If405203c72742c03e73de6f816570b8156218108