FlowState: introduce byte buffer block streaming
This patch introduces an input and output streams that use a queue
of byte buffers (blocks) as the underlying storage mechanism. Blocks are
added on-demand to the queue from a pool of finite byte buffers. Each
block contains a header specifying at least the length of the data
written. This patch also introduces an expiration mechanism for the
blocks allocated to a given stream. After a specified amount of time,
blocks are removed from the head of the queue and returned to the pool
of byte buffers to be later reused. This components will be used in
subsequent patches to implement a memory mapped ring buffer to store
flow state.

Ref.: MI-985
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: If405203c72742c03e73de6f816570b8156218108