FlowState: provide on-demand allocation in object pools
Flow state will be compressed and stored in blocks of memory mapped
files. As flow state expires after a configured amount of time, so
should the blocks storing that flow state (ring buffer made of blocks).

The current implementation of the pool pre-allocates all the objects in
the pool. This can be overkill if we set a high limit of flow storage
per port (as each port will pre-allocate the space in disk and the page
cache as soon as the pool is created).

It seems reasonable to dynamically grow the pool on-demand when the
stream of flow state requires it. Reusing blocks of a given file
in the same pool should prevent the file from growing indefinitely
while providing an streaming interface where continuously write to.

This patch introduces a new ObjectPool that allocates the objects
on-demand instead of pre-allocating them. It also provides a release
method that sets the unused objects to null to release any memory
references to those objects. The user of the pool should be responsible
to make any other operations to release that object (e.g. close without
flushing that memory mapped byte buffer in the case of our flow state
implementation).

Ref.: MI-985
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: I3be209ef79b1afffb6c7605280ac386f25daf470