10kp: A state entry decoder
This introduces a decoder trait to use with the state proxy
service. The decoder parses key/value strings retrieved from
storage into KeyValue messages that are serialized and sent to
clients.

The patch provides optimized decoders for the following types:
MAC addresses, IPv4 addresses and UUIDs. All other types are
decoded using the UTF-8 representation of the character string.

Ref: MI-999

Change-Id: I9cd1a805d7003a2c32151452b71b9132faff5e38
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>