10kp: A state table observer
This provides an interface for a state table observer. This trait
merges all customary RX notifications (next, error, complete) in
a single method that receives a Notify protocol buffers message
instead.

The method returns a future which returns a future that completes
when the observer has finished processing the notification allowing
the sender to throttle the rate of notifications.

Ref: MI-999

Change-Id: Ie239e648daadd1e19cf58dd2fae9ea4494c91524
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>