Update patch set 12
Patch Set 12:

We cannot change the name in the protobuf for the existing fields, due to a short sighted decision to use TextFormat to store the data.

However, I'm more concerned with the ambiguity in the logic part of the code. In Port, there should be 4 fields, 

preFilterInboundMirror, postFilterInboundMirror, preFilterOutboundMirror and postFilterOutboundMirror.

There should only be one mirror trait, which handles all 4 mirror positions.

Patch-set: 12