10kp: A state table cache
A state table cache is the server-side implementation of a
legacy replicated map. It has two functions: (i) to synchronize
with the NSDB backend the key-value entries for a particular
state table, and (ii) to accept the subscriptions of observers
that receive notifications with the changes in the table.

A state table cache is identified by the same arguments that
identify a state table: the object class, object identifier, key
and value classes, a table name and an optional list of arguments
(e.g. the VLAN ID for bridge MAC tables).

In order to identify the changes in storage for a given table,
the classes caches a local view of the key-value-version entries,
and this changes are subsequently published to subscribers. In
addition, each cached snapshot of the remote table receives a
version number, which with the ZooKeeper backend, is the PZXID of
the table root node (i.e. the ID of the last transaction that
modified the table entries). This version number is exposed to
subscribers, such that in a subsequent iteration of this class
we can enable versioning.

To symplify the synchronizatio across multiple subscriptions,
the state table cache has a linear lifecycle STOPPED-STARTED-CLOSED.
The transition from one state to the next is driven by changes to
the subscripion list, where:

* STOPPED is the initial state with no subscribers when the cache
  does not synchronize with the NSDB backend.

* STARTED is when the cache has one or more subscribers, and it
  maintains watchers that synchronize the internal cache with the
  NSDB backend. This state also maintains a list of the current
  subscriptions.

* CLOSED is either when there are no more subscribers to the cache
  or when an unrecoverable error ocurred while synchronizing with
  the storage. This is a terminal state and at this point no
  further subscriptions are accepted.

Transition between states is atomic, which ensures that a
subscription is either accepted or explicitely rejected when the
cache is closed.

An internal Subscription class implements the StateTableSubscription
trait allowing the caller to control the subscription, and it
encapsulates the corresponding observer. Its purpose is to keep a
per-observer order of updates (e.g. an observer will not receive a
diff until the initial snapshot was published), sanity-check the
diffs (a new diff must match the version of a preceeding snapshot
or diff) as well as ensuring a backpressure mechanism for slow
observers using an unbounded queue.

Ref: MI-999

Change-Id: I75039e73e4dca768f165e015d7ee1827f0b667f1
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>