FlowState: add flow state message streams
This patch introduces the flow state input and output streams. They use
the underlying snappy streams and byte buffer streams to write flow
state data compressed and stored in blocks. These streams make sure to
write the size of the message and the message itself in a single chunk
to avoid being split into separate blocks. For the flow state
encoding/decondig, we make use of the SbeEncoder, a wrapper class that
contains methods to encode/decode information into flow state SBE
messages and provides access the underlying byte array holding the data.

This patch only provides a heap byte buffer allocator so the underlying
storage is not baked by a file yet. This feature will be introduced in
the following patch.

Ref.: MI-985
Change-Id: I136a677c17f22d533fbc1065f967d8d7d5ea3f9e
Signed-off-by: Xavi León <xavi.leon@midokura.com>