Update patch set 3
Patch Set 3: -Code-Review

Just a general comment, not sure if applicable for this branch, can be changed in a later patch or completely out of scope. Right now, we do the binding in different parts of the code (cluster api, agent rest api, and also in service containers). Could we move this binder code to nsdb module (or utils) and make all binding code paths converge into a single piece of code?

Patch-set: 3
Label: Code-Review=0