FlowState: Add previous owner field
Add support for a previous_host_id in NSDB so that there is visibility
on the host that used to own the flowstate of a port. With this new
information it is possible to decide if a given flowstate request should
be made to local storage or to a remote one.

Furthermore, in the case of a remote request, this new field is used to
tell the Flow State Minion where to go find the flow state requested.

Ref: MI-988

Change-Id: Iabb6932d7b5613c7a23144b23091e9e2d37ff8f9
Signed-off-by: Mateo Hermida <mateo@midokura.com>