Update patch set 5
Patch Set 5:

> (1 comment)

Calling it rest_api or a generic name implies that the agent has a generic rest api, for any use. It's not. We shouldn't even need this rest api. It's a hack to workaround a limitation neutron is putting on us. Rest api's should go in the cluster.

Patch-set: 5
Label: Code-Review=0