midolman: fix openvswitch-datapath-dkms dependency
It was intended, by having newer openvswitch-datapath-dkms conflict on
newer releases and have the package optionally depend otherwise on
newish linux images, that result would be:

* Install a newish linux-image that includes openvswitch.ko
* If none is available, install openvswitch-datapath-dkms

I suspect that this is a leftover from Ubuntu 12.04. In any case,
the order in which dependencies are processed is to take the first one
that is valid (for some reason the depends and conflicts at the same
time was not evaluated or a match for an older openvswitch-datapath-dkms
was found.

The solution I decided on taking was to still allow
openvswitch-datapath-dkms, but only as a last resort if there is no new
image. The risk of this is that if somebody has a custom image package
that does not match, it may pull the installation of dkms unintendedly.

Fixes: MI-1102
Change-Id: I524b862b85682e6b688dd021f763c30c597f1152
Signed-off-by: Antoni S. Puimedon <antonisp@celebdor.com>