FlowState: WIP for removal of Cassandra
This WIP patch introduces several changes:

A new Flow State Protocol using Protocol Buffers, so that we can have an
extensible communication protocol between different agets and flow state
minions. This is a key part in the exchange of flow state, since we need
them to orchestrate it.

Changes in the flow state minion so that it can add a new TCP frontend
form where to listen for the new Flow State Protocol transfers.

Changes in the MidoNet agent, so that it can take a flag to decide
whether to read/write flowstate from the legacy Cassandra cluster or from
other agents. This switch will allow for a smoother change.

Added a previousOwner field to the ports, so that an agent can know
where to go look for a flow state after a migration.

Ref: MI-987

Change-Id: Icee6b7a434c9399c3fd3fd42ff459f12d8559ab0
Signed-off-by: Mateo Hermida <mateo@midokura.com>