FlowState: builders for flow state writers/readers
This patch introduces the main interfaces for the flow state message
output and input streams as well as companion objects to instantiate
those streams (along with the raw byte buffer block streams) using
the provided configuration and corresponding port id.

This patch allows other patches to build on top of it without worrying
about the underalying implementation details.

Ref.: MI-985
Change-Id: I21f8631fed25279ff12b4c528fc6be87fa1f6941
Signed-off-by: Xavi León <xavi.leon@midokura.com>