10kp: A state table subscriber list
This class manages the current subscriber list at the server for
all state tables. The class maintains a dual index: state table
key to subscriber and subscription identifier to subscriber, where
changes to any index are atomically reflected in the other.

The list allows fetching a subscriber by state table key, or
removing a subscriber by subscription identifier. All changes to
the subscription list are lock free.

Since the getOrElseUpdate method may have side effects (e.g.
creating a new subscriber when concurrent calls modify the same
subscriber table), the method accepts functions that allow the
caller to undo the side effects (i.e. close the unused subscription,
as well as to determine where the subscription existed or is new).

Ref: MI-999

Change-Id: I6ed421d4340083a5f2f67e7aa17e065caf7ad0d7
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>