FlowState: Add TCP client for intercommunication
Adds a new TCP client that will be used to request Flow State from the
Flow State Minions exposing it. The same client can be used for the
three new type of requests (raw, internal and remote).

Also added a mechanism to allow for pipelined encoding of the flow state
that is being read for performance.

Ref: MI-987

Change-Id: I8fd18ecc420ea52d34a5f7b0fa0494a33db5d02f
Signed-off-by: Mateo Hermida <mateo@midokura.com>