BGP-IP: don't restart bgpd on reconfiguring
The restart of bgpd when reconfiguring was necessary for 2 reasons:

1) We needed to reconfigure the devices that bgpd was thought to
   bind to, in the namespace hosting the daemon.
2) We did not handle the "next hop lookup" bgpd request, which came
   after a restart.

The first reason turned out to not be necessary, after some testing.
bgpd binds to 0.0.0.0:179, and reconfiguring individual namespaces
does not break this connection.

The second reason is fixed in this patch. This allows us to not
restart bgpd at all when reconfiguring.

Change-Id: I869f4636bea9727d33e00c23deeec2dc512645d9