FlowState: Mechanism to figure out previous owner
Adds a mechanism to allow a MidoNet agent to figure out who was the previous
owner of a particular port given a portId. For this we need to add a new
previousHostId so that we have a reference to who had the flow state for a
given port.

Adds changes to the two parts where the previous host needs to be updated
during bind/unbind operations. Here it is important to differenciate
between active and non active ports. If the port is active, then it is
the actual owner that holds the current flowstate, otherwise it is the
previous owner.

Ref: MI-988

Change-Id: I6a36ccd4f63c921e2e7c1b807cdcaa0e1860d2e1
Signed-off-by: Mateo Hermida <mateo@midokura.com>