Logging: Add logger with marker wrapper
This adds a Logger wrapper that includes an optional marker to
each log message. This will be used to substitute individual
loggers in the agent/cluster per device or ZK node with common
loggers that include a custom marker.

The rationale for this is to prevent caching of an increasing
number of loggers in the logback LoggingContext while minimizing
the number of code changes: we shall use the new Logger instead
and each class can add its own mark prefix.

The wrapper uses logging markers rather than simply prefixing a
string, to prevent the generation of additional string garbage
for each logging messages, as well as to allow the operator to
filter the messages by marker, and modify the wrapper formatting
as needed.

Ref: MI-1075

Change-Id: I55f4939f5c47f849b939c54ee82a8a2769961e00
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>