Fix race in ReplicatedMapStateTableTest
This fixes an issue in ReplicatedMapStateTableTest where a race
condition between adding a new value for an existing key in the
map and removing the old value would cause the test to fail.

This happens because the test performs a remove of the key from
the table, under the assumption that at that point the table
already owns the version for the latest value of that key.
However, becuase of the aforementioned race condition, it it
possible that the owned versions has not yet been updated at that
point.

The fix is to retry the remove operation until it succeeds.

Change-Id: Id8bc47ddd8c5ca5f0140b065e6d946c8e7b5532e
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>