FlowState: add clearing and read only factory for mem mapped blocks
References to the memory mapped buffers on a given file may
be held undefinitely by the underlying ring buffer. This patch adds
methods to remove the references held by the ring buffer.

This patch also provides a read only factory to open memory mapped
blocks in read only mode without initializing block headers. This
will be used in later patches to scan the flow state file to find the
actual head and tail of the ring buffer.

Ref.: MI-985

Change-Id: Idca2afedc70444cce308871654d373a9c64a800c
Signed-off-by: Xavi León <xavi.leon@midokura.com>