FlowState: send incomming flow state to minion
This patch enabled the agent to send incoming flow state from other
Midonet Agents to the flow state minion. This is only necessary when
storing flow state messages locally. In this case, we need to know which
ports are owned by the agent in order to store the message with the
corresponding writer indexed by port ID.

To support this use case, we extend the minion UDP interface to also
receive a message with the set of ports currently bound locally, and
intersect this set with the port ids present in the message.

Another possibility would be to extend the flow state SBE to include
the currently owned ports but it was discarded because it meant
regenerating the whole message to include the owned ports field,
increasing the number of memory allocations in the fast path.

Change-Id: I5c60d48c1eceb69654e50541873aafc0186f2c4e
Signed-off-by: Xavi León <xavi.leon@midokura.com>