Update patch set 4
Patch Set 4:

So the macro itself doesn't affect the memory usage here, this is the same as the macro we're using from scalalogging. The reason for the macro is to defer the evaluation of the macro arguments if a certain logging level is not enabled (this is done by scalalogging). E.g.

log.debug(s"Result: ${callExpensiveFunction()}")

the callExpensiveFunction() is always evaluated if debug would be a method,  but it will not if debug is a macro and debug is disabled.

Anyway this patch only adds the marker to include as a prefix, this it to prevent creation of loggers, such as:

org.midonet.devices.bridge.bridge-some-id

where we create a new logger for every device and not gc them. However the macros are needed for the reason above.

Patch-set: 4
Label: Code-Review=0