10kp: Server client context
A client context manages all state table subscriptions for a
given client, and enables the setup of new subscriptions or
teardown of existing ones, to be notified back to a remote client
with a message sent via the client handler.

The context manages all subscriptions using the state table
subscriber list, and handles concurrent access by cleaning up the
unused subscriptions.

The context also allows to gracefully teardown the existing
subscription when the server shuts down and notify the remote
clients.

Ref: MI-999

Change-Id: I8f190b3273c2ae355f2cdb92e78b998f2693ebbf
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>