Add HdrHistogramSlidingWindow reservoir for metrics
This allows us to specify a sliding window over which we collect latency
information. For example, we can specify that we want the last 5 minutes
worth of latency information, which can then be polled every 1 minute by
monitoring software to see how latency percentiles change over time.

We store the latencies using hdr histogram for accuracy.

Ref MI-955

Change-Id: I096a94e1b7c78c13c83a6ee5c8578a4a826e086f
Signed-off-by: Ivan Kelly <ivan@midokura.com>