10kp: State proxy server handles requests
This patch integrates the state proxy manager with the Netty
message handler, allowing the client to respond to incoming
requests. The following events are handled by the manager:

* Registering a new client channel.

* Unregistering an existing client channel.

* A SUBSCRIBE request for a state table.

* An UNSUBSCRIBE request for a previous subscription.

Responses for these events are returned by the manager, which
uses the corresponding client handler to send the response
message to the client channel. Also the manager will dispatch
the subscriptions notifications from the state table cache
corresponding to each client subscription. The manager will
handle cleanup messages during server shutdown, as well as
subscription cleanup for connections closed remotely by the
clients.

In addition, the server protocol handler responds immediately to
a client in the following situations (these requests are not
handled by the state proxy manager):

* A PING request.

* Unknown messages or channel exceptions.

* Any exception thrown when calling the manager.

Ref: MI-999

Change-Id: Ia90779961a760175afe5e2f6e10ab834d8f7afb2
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>