Fix MTU DHCP test
The MTU DHCP test verifies that the MTU value from an DHCP reply
is the same as the one configured in the simulation. In our DHCP
implementation we use the MTU value set in the DHCP subnet
configuration (in NSDB), however the test compares the MTU value
with the minimum MTU configured in the datapath controller. This
latter value is static, mutable and modified by a previous test
to 4200 from the default 1450. Therefore, whenever the datapath
controller MTU doesn't match the expected one, the test would fail.

This changes the test MTU value to 1450, which is the value used
by the topology builder when creating a new DHCP subnet.

Change-Id: I9100f5e7e72744d7213e072d124010848eea92e2
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit 2c54191e8975d6ec4c08c6f033f56e2dbb973300)