10kp: Add client connection state
This modifies the state table client trait to add a connection
observable. This observable will emit upon subscription a
notification with whether the client is connected to a proxy
state server, and thereafter a notification whenever the connection
state changes.

The connection obervable may only terminate when the client stops.

Ref: MI-1003

Change-Id: Icdaaf54373486be7c24155f3a720b1dd33e662e7
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>