Use the actual interface MTU for tunnel interfaces
Right now, there were several issues on how we specify the DHCP MTU
option to vms:
- we were using the configuration value of the DHCP MTU to compute
  the minimum MTU from the tunnel interfaces, instead of the actual
  MTU of that tunnel interface.
- the logic to compute the MTU to send to vms was spread on several
  places (DatapathController, PacketWorkflow and DHCPImpl)

This patch fixes how the underlay MTU is computed and sent to vms
through DHCP. The main functionality fixes are: i) use the actual
MTU of the tunnel interfaces instead of the configured value;
and ii) get the minimum of the bridge MTU value and the underlay
MTU.

This patch also attempts to simplify how the DHCP MTU option is
computed by centralizing the logic on the DHCPImpl class. It also
decouples the underlay MTU discovery from the DHCP MTU configuration
option so the DatapathController always reports the minimum MTU among
the tunnel interfaces.

Ref.: MI-1129
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: I926cc3d040d9613847ebafe33d3b9a602059ddaa