Workaround to fix races during MAC learning
Syseleven is experiencing a high number of concurrent writes when
learning MAC entries. According to the logs, several entries with the
same mapping port-MAC are updated concurrently. Instead of just adding
the first one to zookeeper, it looks like we don't cover some race
conditions which implies writting several equal entries to zookeeper
growing the replicated map indefenitely, breaking the zookeeper
connection because the size of the directory is too big.
This is a hotfix to stabilize the current situation which locks the
TimedExpirationMap to avoid concurrent modifications. A proper fix should
entail finding where the race condition happen and provide the
appropiate non-blocking implementation.

Change-Id: I3826bbc83c6e428308829c9b4f0698cdb0d36fa9
Signed-off-by: Xavi León <xavi.leon@midokura.com>