Prevent adding duplicate entries to replicated map
This verifies that when adding a new entry to a replicated map,
the same key-value pair is not in the process of already being
added.

This branch was overwritten. The previous commit-id that can be
found on midolman logs is 428eab97254d7d7a2481f1d7e3ac1d0fbf726684

Ref: MI-1068

Change-Id: I8c93903c8292eeda53284704826ede1815fbc5d7
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>