BGP-IP: use router id in namespace name
The current naming scheme for bgp on exterior ports
uses the index id to make the namespace names unique.
This works on a per-host basis but does not uniquely
identify hosts across the entire deployment.

Adding the router-id to the namespace name makes the
namespace name unique across deployments and has the
added benefit of letting the operator know which router
the namespace is associated with.

Change-Id: I047842e14ec25fd6b9dea24d6e5b1580362bcc93