FlowState: Add port migration test
The flow state transfer on a migration is now handled by MidoNet using
an internal storage instead of Cassandra. This test is used for the
exchange that will be made between agents once there is a port migration
between different hosts. This test uses the default Agent configuration
(fetch/push state from/to Cassandra) and using the two current
mechanisms for bind/unbind (i.e. using the Midonet API and using the
mm-ctl tool).

Ref: MI-1168

Change-Id: I74104a65af45b4e6df6b2941b85d57ecdf8369b9
Signed-off-by: Mateo Hermida <mateo@midokura.com>
Signed-off-by: Xavi León <xavi.leon@midokura.com>