Do not restart workflow for processed packets
Currently if the future for a postpone operation of a packet
completes after the packet was processed we restart the workflow.
However, since tha packet context was released back to the
context pool, restarting the workflow will generate a null pointer
exception.

This patch prevents restarting the workflow for a context that
was returned to the context pool by verifying the packet context
cookie.

Fixes: MI-1166

Change-Id: I02139c1166f11e80e4b2033009ae97f513677c86
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>