Cache tunnel interfaces so we update the MTU on a tunnel zone update
The underlay MTU is not updated whenever we receive an update on the
tunnel zones. This can make the MTU outdated if the interfaces are added
or removed from tunnel zones.

This patch caches the interfaces when the interface scanner sends an
update which are later used to compute the MTU of the tunnel interfaces
added to the host.

Ref.: MI-1129
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: I52c90caf92007adb94ab1d84a74648c32f925af7