Wait for tunnel zones before initialization of datapath controller
There is a bug in the initialization sequence of the datapath controller
by which we may update the host interfaces as well as the underlay MTU
before the tunnel zones are received. In this case, the underlay MTU is
incorrectly computed because the current set of interfaces in a tunnel
zone is empty (we pick the default). We don't recompute the correct
underlay MTU until the there's a new interface change triggered by the
interface scanner.

This patch makes sure we don't set the state of the datapath controller
as initialized until we have all necessary information, i.e. tunnel
ports created, tunnel zones received from storage and the initial set of
interfaces received from the interface scanner. Additionally, we cache
the latest interface update received so when we receive an update on the
tunnel interfaces, we are able to recompute the corresponding minimum
MTU.

While at it, avoid calling twice on a method when we already cache the
value locally on the DefaultInterfaceScanner.

Ref.: MI-1129
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: I52c90caf92007adb94ab1d84a74648c32f925af7