FWLOG: Binary FWaaS logging
Switch RuleLogger from text logging with slf4j and Logback to binary logging
using SBE and a BufferedOutputStream. Also add a deserializer that reads the
files. The latter is used only for testing in this patch, but will be used for
the log translation tool in a later patch.

Remove file_name property from RuleLogger due to concerns about moving the disk
head when writing to multiple log files. Now all events will be written to a
single binary log file, with separation done in the translation tool.

Bounds checking is not yet implemented. If the total binary-encoded size of the
event exceeds 8k, logging will fail. Since metadata is the only field with
arbitrary length and is not exposed in the API, and the FirewallLogTranslator
doesn't create metadata anywhere close to that size, this shouldn't be a
problem, but when we expose metadata via the API, we'll need to enforce limits.

Change-Id: I9a5d06cd16d34cd30b2298451920f4f1557157d0
Signed-off-by: Brandon Berg <bberg@midokura.com>